arm: dts: Resync dtsi files with kernel 4.19 from Beacon
The baseboard and SOM dtsi files are not in sync with the kernel,
with a few exceptions where nodes were deleted due to the entries
missing from the SoC DTSI file in U-Boot.

A modification was made to r8a774a1-beacon-rzg2m-kit-u-boot.dtsi
to diable the sdhi controller with the Wi-Fi controller.  U-Boot
does not use the WiFi and removing it avoids a regression,
because the WiFi controller would change the MMC index.

Signed-off-by: Adam Ford <aford173@gmail.com>