configs: beacon-rzg2m.h: Fix loading addresses
The default settings for loadaddr work with newer kernels, but fail
to boot the Beacon master kernel (4.19) due to the location the
kernel is being loaded.

This patch moves the fdt_addr and loadaddr back to their defaults
of 0x48080000 and 0x48000000 resepctively which restores booting
from the older kernel.

Signed-off-by: Adam Ford <aford173@gmail.com>