arm64: beacon-rzg2m: Enable Auto-detection of M/N/H variants
Currently, the config file specifies the device tree file,but it
may be possible to use one U-Boot with all three boards.

This patch removes the fdt_file definition from beacon-rzg2m.h and
a late initialization function checks the processor type and sets
the fdt_file based on the SoC type.

Signed-off-by: Adam Ford <aford173@gmail.com>