ARM: rmobile: beacon: Change environment offset
The environment offset had been set to 0 placing it at the
beginning of the second hardware boot partion. This conflicts with
the boot ROM use of the second partition for redundent boot. Place
at the end of partition 2 by setting offset to a "negative" number.

Also change the environment size to 8K. It had defaulted to just short
of 128K. NOTE: the default environment is less than 2K.

Signed-off-by: Charles Stevens <charles.stevens@compasses.com>